% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_Celaref}
\alias{AutoAnnotate_Celaref}
\title{Automatic annotation of Seurat object using Celaref package.}
\usage{
AutoAnnotate_Celaref(SeuratObj, counts_ref, cellinfo_ref, plot = TRUE)
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{counts_ref}{counts matrix of reference data}

\item{cellinfo_ref}{data frame of cell information of reference data, the first column as cell ID,
the second column as cluster ("Cluster" as column name)}

\item{plot}{produce plots or not}
}
\value{
data frame of prediction result
}
\description{
Automatic annotation of Seurat object using Celaref package.
}
