% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_CellAssign}
\alias{AutoAnnotate_CellAssign}
\title{Automatic annotation of Seurat object using CellAssign package.}
\usage{
AutoAnnotate_CellAssign(SeuratObj, markerDF)
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{markerDF}{A data frame with two columns (celltype, gene), containing the markers of each cell type.}
}
\value{
Seurat object. The predicted cell type label is stored in \code{SeuratObj$predicted_label}.
}
\description{
Automatic annotation of Seurat object using CellAssign package.
}
