% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_scCATCH}
\alias{AutoAnnotate_scCATCH}
\title{Automatic annotation of Seurat object using scCATCH package.}
\usage{
AutoAnnotate_scCATCH(SeuratObj, marker_custom)
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{marker_custom}{Please refer to \url{https://raw.githack.com/ZJUFanLab/scCATCH/master/vignettes/tutorial.html} to build a marker data.frame}
}
\value{
data frame of prediction result
}
\description{
Automatic annotation of Seurat object using scCATCH package.
}
