% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_SingleR}
\alias{AutoAnnotate_SingleR}
\title{Automatic annotation of Seurat object using SingleR package.}
\usage{
AutoAnnotate_SingleR(SeuratObj, ref, ref_type = c("bulk", "single-cell"))
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{ref}{Reference dataset, bulk RNA-seq or microarray data or single-cell data,
\code{SummarizedExperiment} or \code{SingleCellExperiment} object containing a matrix of log-expression values
with labels stored as \code{ref$label}.}

\item{ref_type}{type of reference dataset, \code{"bulk"} (bulk RNA-seq) or \code{"single-cell"} (single-cell data) are supported.}
}
\value{
Seurat object. The predicted cell type label is stored in \code{SeuratObj$predicted_label}.
}
\description{
Predict the cell type of Seurat object using SingleR package.
}
