% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIEX_visualization.R
\name{BordaRank_MINIEX}
\alias{BordaRank_MINIEX}
\title{Dot plot showing top regulons of each cluster according to Borda ranking}
\usage{
BordaRank_MINIEX(MINIEXouputPath, cluster, topn = 10)
}
\arguments{
\item{MINIEXouputPath}{output path of MINI-EX.}

\item{cluster}{which cluster to show}

\item{topn}{number of top regulons to draw}
}
\description{
Dot plot showing top regulons of each cluster according to Borda ranking
}
\examples{
\dontrun{
BordaRank_MINIEX(MINIEXouputPath="regulons_output", cluster=1, topn = 10)
}

}
