% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constellationPlot.R
\name{constellationPlot}
\alias{constellationPlot}
\title{Draw Constellation plot with scrattch.hicat package.}
\usage{
constellationPlot(
  SeuratObj,
  cluster_label = "cell_type",
  out.dir = "Constellation_plots",
  k = 15,
  knn.outlier.th = 2,
  outlier.frac.th = 0.5
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{cluster_label}{one column of \code{SeuratObj@meta.data}, used for cluster annotation to display in the plot.}

\item{out.dir}{location to write plotting files to}

\item{k}{K for KNN algorithm.}

\item{knn.outlier.th}{Threshold to determine if a nearest neighbor is too far}

\item{outlier.frac.th}{Threshold to determine if a cell is a outlier if it is too far away from most of its neighbors.}
}
\description{
Draw Constellation plot with scrattch.hicat package.
}
\examples{
\dontrun{
SeuratObj <- pbmc3k.SeuratData::pbmc3k.final
Seurat::Idents(SeuratObj) <- SeuratObj$seurat_clusters
constellationPlot(SeuratObj, cluster_label = "seurat_annotations", out.dir = "Constellation_plots")
}

}
