% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeBulk.R
\name{decomposeBulk}
\alias{decomposeBulk}
\title{Decompose bulk RNA-seq samples using a single cell reference}
\usage{
decomposeBulk(
  SeuratObj,
  exprs,
  meta,
  mc.cores = 1,
  slot = "counts",
  assay = "SCT"
)
}
\arguments{
\item{SeuratObj}{A Seurat object.}

\item{exprs}{A data frame or matrix of raw read counts of \epmh{bulk} RNA-seq samples. Column names correspond to sample names, row names to genes.}

\item{meta}{A data frame of 2 columns with Sample and Class, correlate sample names with sample class.}

\item{mc.cores}{Number of cores used}

\item{slot}{Which slot to use to calculate mean expression per cell type}

\item{assay}{Which assays to use to calculate mean expression per cell type}
}
\value{
A data frame in long format, indicating the fraction of cell types in each bulk RNA-seq sample.
}
\description{
CIBERSORT is an algorithm (Newman et al. Nautre Methods 12:453-457) for estimating the cell type composition of a bulk sample,
given a gene expression profile of the sample and a known gene expression profile for each cell type potentially contributing to the sample.
Here, we wrap runCIBERSORT function from RNAMagnet R package to decompose bulk RNA-seq samples.
}
