% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIEX_visualization.R
\name{enrich_exp_hmp}
\alias{enrich_exp_hmp}
\title{Heatmaps showing cluster enrichment and TF expression of each regulon in each cluster.}
\usage{
enrich_exp_hmp(
  SeuratObj,
  MINIEXouputPath,
  group.by = "seurat_clusters",
  assay = "SCT"
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{MINIEXouputPath}{output path of MINI-EX.}

\item{group.by}{Name of the metadata column to group cells by (for example, seurat_clusters)}

\item{assay}{assay to calculate mean expression}
}
\description{
Heatmaps showing cluster enrichment and TF expression of each regulon in each cluster.
}
\examples{
\dontrun{
enrich_exp_hmp(SeuratObj, MINIEXouputPath="regulons_output")
}

}
