% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIEX_visualization.R
\name{enrich_exp_scatter}
\alias{enrich_exp_scatter}
\title{Dimension reduction plot showing cluster enrichment and TF expression}
\usage{
enrich_exp_scatter(SeuratObj, MINIEXouputPath, gene, reduction = "umap")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{MINIEXouputPath}{output path of MINI-EX.}

\item{gene}{TF to show}

\item{reduction}{umap or tsne}
}
\description{
Dimension reduction plot showing cluster enrichment and TF expression
}
\examples{
\dontrun{
enrich_exp_scatter(SeuratObj, MINIEXouputPath="regulons_output", gene = 'AT2G40750')
}

}
