% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Ath}
\alias{example_Ath}
\title{Toy example single-cell expression matrix of Arabidopsis thaliana.}
\format{
A dgCMatrix of 7000×2000(genes×cells).
}
\usage{
example_Ath
}
\description{
7000×2000(genes×cells) single-cell expression matrix of Arabidopsis thaliana, taken as a toy example.
}
\keyword{datasets}
