% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Osa}
\alias{example_Osa}
\title{Toy example single-cell expression matrix of Oryza sativa.}
\format{
A dgCMatrix of 7000×2000(genes×cells).
}
\usage{
example_Osa
}
\description{
7000×2000(genes×cells) single-cell expression matrix of Oryza sativa, taken as a toy example.
}
\keyword{datasets}
