% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Zma}
\alias{example_Zma}
\title{Toy example single-cell expression matrix of Zea mays.}
\format{
A dgCMatrix of 7000×1000(genes×cells).
}
\usage{
example_Zma
}
\description{
7000×1000(genes×cells) single-cell expression matrix of Zea mays, taken as a toy example.
}
\keyword{datasets}
