% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeBulk.R
\name{fractionHmp}
\alias{fractionHmp}
\title{Heatmap showing the fraction of cell types in each bulk RNA-seq sample.}
\usage{
fractionHmp(CIBER, meta)
}
\arguments{
\item{CIBER}{output of \code{decomposeBulk()}, a data frame indicating the fraction of cell types in each bulk RNA-seq sample.}

\item{meta}{A data frame of 2 columns with Sample and Class, correlate sample names with sample class.}
}
\description{
Heatmap showing the fraction of cell types in each bulk RNA-seq sample.
}
