% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{get_marker_file}
\alias{get_marker_file}
\title{Generate a marker file to run Garnett.}
\usage{
get_marker_file(markerDF, SeuratObj, file = "./markers.txt")
}
\arguments{
\item{markerDF}{A data frame with two columns (celltype, gene), containing the markers of each cell type.}

\item{SeuratObj}{The Seurat object that you going to annotate using Garnett.}

\item{file}{Path to the marker file to be generated. Default is "./markers.txt".}
}
\value{
None, marker file is written to \code{file} parameter location.
}
\description{
Generate a marker file to run Garnett.
}
