% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{get_marker_scCATCH}
\alias{get_marker_scCATCH}
\title{Generate a data frame of marker genes with the required format to run scCATCH.}
\usage{
get_marker_scCATCH(markerDF, species, tissue)
}
\arguments{
\item{markerDF}{A data frame with two columns (celltype, gene), containing the markers of each cell type.}

\item{species}{species.}

\item{tissue}{tissue.}
}
\value{
A data frame containing the markers of each cell type.
}
\description{
Generate a data frame of marker genes with the required format to run scCATCH.
}
