% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_rice_IRGSP_1_0}
\alias{GO_rice_IRGSP_1_0}
\title{GO annotation of Oryza sativa (IRGSP-1.0).}
\format{
A data frame with 175728 rows and 3 variables:
\describe{
\item{Gene.stable.ID}{gene id of rice}
\item{GO.term.accession}{accession number of GO terms}
\item{GO.term.name}{names of GO terms}
}
}
\source{
\url{https://plants.ensembl.org/index.html}
}
\usage{
GO_rice_IRGSP_1_0
}
\description{
A dataset containing rice genes and their GO annotation.
}
\keyword{internal}
