% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_rice_V7}
\alias{GO_rice_V7}
\title{GO annotation of Oryza sativa (v7.0).}
\format{
A data frame with 70633  rows and 3 variables:
\describe{
\item{Gene.Name}{gene id of rice}
\item{GO.ID}{accession number of GO terms}
\item{GO.Description}{names of GO terms}
}
}
\source{
\url{https://phytozome-next.jgi.doe.gov/}
}
\usage{
GO_rice_V7
}
\description{
A dataset containing rice genes and their GO annotation.
}
\keyword{internal}
