% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_zeamays_NAM_5}
\alias{GO_zeamays_NAM_5}
\title{GO annotation of Zea mays (Zm-B73-REFERENCE-NAM-5.0).}
\format{
A data frame with 291353 rows and 3 variables:
\describe{
\item{Gene.stable.ID}{gene id of Zea mays}
\item{GO.term.accession}{accession number of GO terms}
\item{GO.term.name}{names of GO terms}
}
}
\source{
\url{https://plants.ensembl.org/index.html}
}
\usage{
GO_zeamays_NAM_5
}
\description{
A dataset containing Zea mays genes and their GO annotation.
}
\keyword{internal}
