% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GO_zeamays_V4}
\alias{GO_zeamays_V4}
\title{GO annotation of Zea mays (RefGen_V4).}
\format{
A data frame with 233821 rows and 3 variables:
\describe{
\item{Gene.Name}{gene id of Zea mays}
\item{GO.ID}{accession number of GO terms}
\item{GO.Description}{names of GO terms}
}
}
\source{
\url{https://phytozome-next.jgi.doe.gov/}
}
\usage{
GO_zeamays_V4
}
\description{
A dataset containing Zea mays genes and their GO annotation.
}
\keyword{internal}
