% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_rice_V7}
\alias{KEGG_rice_V7}
\title{KEGG annotation of Oryza sativa (v7.0).}
\format{
A data frame with 10186  rows and 3 variables:
\describe{
\item{Gene.Name}{gene id of rice}
\item{KEGG.ID}{accession number of KEGG terms}
\item{KEGG.Description}{names of KEGG terms}
}
}
\source{
\url{https://phytozome-next.jgi.doe.gov/}
}
\usage{
KEGG_rice_V7
}
\description{
A dataset containing rice genes and their KEGG annotation.
}
\keyword{internal}
