% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KEGG_zeamays_NAM_5}
\alias{KEGG_zeamays_NAM_5}
\title{KEGG annotation of Zea mays (Zm-B73-REFERENCE-NAM-5.0).}
\format{
A data frame with 16073  rows and 3 variables:
\describe{
\item{Gene.Name}{gene id of Zea mays}
\item{KEGG.ID}{accession number of KEGG terms}
\item{KEGG.Description}{names of KEGG terms}
}
}
\source{
\url{https://phytozome-next.jgi.doe.gov/}
}
\usage{
KEGG_zeamays_NAM_5
}
\description{
A dataset containing Zea mays genes and their KEGG annotation.
}
\keyword{internal}
