% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif2TF_At}
\alias{motif2TF_At}
\title{Correspondence between motif and transcription factor in Arabidopsis thaliana.}
\format{
A named vector of 1366 character.
}
\usage{
motif2TF_At
}
\description{
Correspondence between motif and transcription factor in Arabidopsis thaliana.
}
\keyword{internal}
