% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif2TF_Os}
\alias{motif2TF_Os}
\title{Correspondence between motif and transcription factor in Oryza sativa.}
\format{
A named vector of 476 character.
}
\usage{
motif2TF_Os
}
\description{
Correspondence between motif and transcription factor in Oryza sativa.
}
\keyword{internal}
