% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif2TF_Zm}
\alias{motif2TF_Zm}
\title{Correspondence between motif and transcription factor in Zea mays.}
\format{
A named vector of 505 character.
}
\usage{
motif2TF_Zm
}
\description{
Correspondence between motif and transcription factor in Zea mays.
}
\keyword{internal}
