% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orthologs_At_Os}
\alias{orthologs_At_Os}
\title{One-by-one orthologous genes of rice in Arabidopsis thaliana.}
\format{
A named vector of 12013 character.
}
\usage{
orthologs_At_Os
}
\description{
One-by-one orthologous genes of rice in Arabidopsis thaliana, used for cross-species scRNA-seq analysis.
}
\keyword{internal}
