% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orthologs_At_Zm}
\alias{orthologs_At_Zm}
\title{One-by-one orthologous genes of Zea mays in Arabidopsis thaliana.}
\format{
A named vector of 11405 character.
}
\usage{
orthologs_At_Zm
}
\description{
One-by-one orthologous genes of Zea mays in Arabidopsis thaliana, used for cross-species scRNA-seq analysis.
}
\keyword{internal}
