% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMET_analysis.R
\name{PMEThmp}
\alias{PMEThmp}
\title{Heatmap of PMET result}
\usage{
PMEThmp(PMETresult, topn = NULL, rownames.size = 5, colnames.size = 5)
}
\arguments{
\item{PMETresult}{PMET result}

\item{topn}{show top highly enriched pairs of each cluster}

\item{rownames.size}{font size of row names}

\item{colnames.size}{font size of column names}
}
\description{
Heatmap of PMET result
}
