% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMET_analysis.R
\name{processPMET}
\alias{processPMET}
\title{pre-process original PMET output and filter with PPI information}
\usage{
processPMET(PMETresult, PPIinfo = NULL, species = "Arabidopsis thaliana")
}
\arguments{
\item{PMETresult}{original PMET output}

\item{PPIinfo}{protein-protein interaction (PPI) information}

\item{species}{species; currently \code{Arabidopsis thaliana} or \code{Oryza sativa} or \code{Zea mays} are supported.}
}
\description{
pre-process original PMET output and filter with PPI information
}
\examples{
\dontrun{
data(PPIinfo_At)
PMETresult <- read.table(file = "output.txt", header = T, sep = "\t", stringsAsFactors = F)
PMETresult <- processPMET(PMETresult, PPIinfo = PPIinfo_At)
}

}
