% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_visualization.R
\name{ras_exp_scatter}
\alias{ras_exp_scatter}
\title{Dimension reduction plot showing regulon activity and TF expression}
\usage{
ras_exp_scatter(SeuratObj, rasMat, gene, reduction = "umap")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{rasMat}{matrix of regulon activity score(RAS) in each cell}

\item{gene}{TF to show}

\item{reduction}{umap or tsne}
}
\description{
Dimension reduction plot showing regulon activity and TF expression
}
