% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposeBulk.R
\name{runCIBERSORT}
\alias{runCIBERSORT}
\title{modified version of original code(https://rdrr.io/github/veltenlab/rnamagnet/src/R/CIBERSORT.R).}
\usage{
runCIBERSORT(
  exprs,
  base,
  design,
  markergenes = intersect(rownames(base), rownames(exprs)),
  transform = function(x) x,
  nu = c(0.25, 0.5, 0.75),
  optim.nu = F,
  mc.cores = 3,
  ...
)
}
\description{
modified version of original code(https://rdrr.io/github/veltenlab/rnamagnet/src/R/CIBERSORT.R).
}
