% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCytoTRACE.R
\name{RunCytoTRACE}
\alias{RunCytoTRACE}
\title{Run CytoTRACE on Seurat object}
\usage{
RunCytoTRACE(
  SeuratObj,
  keep_orig_reduction = TRUE,
  reduction_method = "umap",
  ncores = 1,
  save_plot = TRUE,
  return_object = TRUE
)
}
\arguments{
\item{SeuratObj}{A Seurat object}

\item{keep_orig_reduction}{keep original reduction dimension of Seurat object.}

\item{reduction_method}{reduction method, umap as default. Only works when \code{keep_orig_reduction = TRUE}.}

\item{ncores}{the number of cores to utilize.}

\item{save_plot}{save plots to pdf or not.}

\item{return_object}{return CytoTRACE result or not.}
}
\value{
CytoTRACE result. See \code{?CytoTRACE::CytoTRACE}.
}
\description{
Run CytoTRACE on Seurat object
}
