% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGSEA_plant.R
\name{RunGSEA_plant}
\alias{RunGSEA_plant}
\title{Perform Gene Set Enrichment Analysis (GSEA) on Seurat object}
\usage{
RunGSEA_plant(SeuratObj, by = "GO", TERM2GENE = NULL, GeneIDtype = "TAIR")
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{GO KEGG. Will be ignored if parameter "TERM2GENE" is not NULL.}

\item{TERM2GENE}{Customized terms. Annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene. NULL by default. If not NULL, parameter "by" will be ignored.}

\item{GeneIDtype}{Gene ID type in Seurat object. \code{TAIR} for \code{Arabidopsis thaliana}, for example.}
}
\value{
Seurat object
}
\description{
Perform Gene Set Enrichment Analysis (GSEA) on Seurat object. GSEA is implemented using \code{clusterProfiler} package.
This function is only for three plant species: \code{Arabidopsis thaliana} or \code{Oryza sativa} or \code{Zea mays}.
}
\examples{
\dontrun{
SeuratObj <- RunGSEA_plant(SeuratObj, by = 'GO', GeneIDtype = 'TAIR')
} 

}
