% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMonocle2.R
\name{RunMonocle2}
\alias{RunMonocle2}
\title{Run monocle2 on Seurat object}
\usage{
RunMonocle2(
  SeuratObj,
  max_components = 2,
  root_clusters = NULL,
  point_size = 0.5,
  BEAM_analysis = FALSE,
  BEAM_branch_point = 1,
  save_plot = TRUE,
  return_object = TRUE
)
}
\arguments{
\item{SeuratObj}{A Seurat object}

\item{max_components}{the dimensionality of the reduced space. 2 as default.}

\item{root_clusters}{clusters selected as root}

\item{point_size}{point size of scatter plot}

\item{BEAM_analysis}{perform BEAM analysis or not.}

\item{BEAM_branch_point}{The ID of the branch point to analyze.}

\item{save_plot}{save plots to pdf or not.}

\item{return_object}{return CellDataSet object or not.}
}
\value{
if parameter \code{return_object = TRUE}, CellDataSet object and differential genes across pseudotime will be returned.
}
\description{
Run monocle2 on Seurat object
}
