% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMonocle3.R
\name{RunMonocle3}
\alias{RunMonocle3}
\title{Run monocle3 on Seurat object}
\usage{
RunMonocle3(
  SeuratObj,
  batch = NULL,
  reduction_method = "UMAP",
  keep_orig_reduction = TRUE,
  root_clusters = NULL,
  save_plot = TRUE,
  return_object = TRUE
)
}
\arguments{
\item{SeuratObj}{A Seurat object}

\item{batch}{one column of \code{SeuratObj@meta.data}, indicating batch effect.}

\item{reduction_method}{reduction method, UMAP as default.}

\item{keep_orig_reduction}{keep original reduction dimension of Seurat object.}

\item{root_clusters}{seurat clusters used to help identify the root.}

\item{save_plot}{save plots to pdf or not.}

\item{return_object}{return CellDataSet object or not.}
}
\value{
if \code{return_object = TRUE}, CellDataSet object, differential genes across pseudotime,
differential genes between clusters, gene modules will be returned.
}
\description{
Run monocle3 on Seurat object
}
