% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSlingshot.R
\name{RunSlingshot}
\alias{RunSlingshot}
\title{Run slingshot on Seurat object}
\usage{
RunSlingshot(
  SeuratObj,
  clusterlabel = "seurat_clusters",
  reduction = "UMAP",
  dynamic_genes = FALSE,
  save_plot = TRUE,
  return_object = TRUE
)
}
\arguments{
\item{SeuratObj}{A Seurat object}

\item{clusterlabel}{cluster labels}

\item{reduction}{reduction}

\item{dynamic_genes}{whether to identify dynamic genes across pseudotime}

\item{save_plot}{save plots to pdf or not.}

\item{return_object}{return object or not.}
}
\value{
if parameter \code{return_object = TRUE}, SingleCellExperiment object and dynamic genes will be returned.
}
\description{
Run slingshot on Seurat object
}
