% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{species_percentage}
\alias{species_percentage}
\title{Bar plot showing the percentage of cells from different species.}
\usage{
species_percentage(SeuratObj, group_by = "seurat_clusters")
}
\arguments{
\item{SeuratObj}{Seurat object.}

\item{group_by}{column of \code{SeuratObj@meta.data}.}
}
\value{
a ggplot ibject.
}
\description{
Bar plot showing the percentage of cells from different species after cross-species integration.
}
