% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_visualization.R
\name{SpecificityRank}
\alias{SpecificityRank}
\title{Dot plot showing top regulons of each cluster according to regulon specificity score(RSS)}
\usage{
SpecificityRank(rssMat, cluster, topn = 10)
}
\arguments{
\item{rssMat}{matrix of regulon specificity score(RSS) in each cluster}

\item{cluster}{which cluster to show}

\item{topn}{number of top regulons to draw}
}
\description{
Dot plot showing top regulons of each cluster according to regulon specificity score(RSS)
}
