% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMET_analysis.R
\name{topPairsNet}
\alias{topPairsNet}
\title{Network diagram showing top pairs of a cluster.}
\usage{
topPairsNet(PMETresult, clus, topn = 20, nodeColor = "#81BADA")
}
\arguments{
\item{PMETresult}{PMET result}

\item{clus}{cluster}

\item{topn}{number of top highly enriched pairs of the cluster to draw, 20 as default.}

\item{nodeColor}{color of nodes}
}
\description{
Network diagram showing top pairs of a cluster.
}
