% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_visualization.R
\name{topRegulons}
\alias{topRegulons}
\title{Network diagram showing top regulons of each cluster}
\usage{
topRegulons(rssMat, topn = 5)
}
\arguments{
\item{rssMat}{matrix of regulon specificity score(RSS) in each cluster}

\item{topn}{number of top regulons to draw}
}
\description{
Network diagram showing top regulons of each cluster according to regulon specificity score(RSS).
}
