% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIEX_visualization.R
\name{topRegulons_MINIEX}
\alias{topRegulons_MINIEX}
\title{Network diagram showing top regulons of each cluster}
\usage{
topRegulons_MINIEX(
  MINIEXouputPath,
  topn = 5,
  from.size = 7,
  to.size = 3,
  from.label.cex = 1,
  to.label.cex = 0.5
)
}
\arguments{
\item{MINIEXouputPath}{output path of MINI-EX.}

\item{topn}{number of top regulons to draw}

\item{from.size}{size of 'cluster' nodes}

\item{to.size}{size of 'regulon' nodes}

\item{from.label.cex}{label size of 'cluster' nodes}

\item{to.label.cex}{label size of 'regulon' nodes}
}
\description{
Network diagram showing top regulons of each cluster according to Borda ranking.
}
\examples{
\dontrun{
topRegulons_MINIEX(MINIEXouputPath="regulons_output", topn = 5)
}

}
