% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_visualization.R
\name{ToptargetsofTopregulons}
\alias{ToptargetsofTopregulons}
\title{Network diagram showing top targets of top regulons of each cluster}
\usage{
ToptargetsofTopregulons(rssMat, tf_target, Topregulons = 5, Toptargets = 5)
}
\arguments{
\item{rssMat}{matrix of regulon specificity score(RSS) in each cluster}

\item{tf_target}{TF and targets of each regulon. This information comes from post-processing of pyscenic.}

\item{Topregulons}{number of top regulons to draw}

\item{Toptargets}{number of top targets to draw}
}
\description{
Each regulator is colored by the cluster that it belongs. Pie means that this regulon is top regulon of multiple clusters.
}
