% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMET_analysis.R
\name{triHmp}
\alias{triHmp}
\title{Draw triangle heatmap of PMET result}
\usage{
triHmp(PMETresult, clus)
}
\arguments{
\item{PMETresult}{PMET result}

\item{clus}{cluster}
}
\description{
Draw triangle heatmap of PMET result
}
