% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_CellAssign}
\alias{AutoAnnotate_CellAssign}
\title{Automatic annotation of Seurat object using CellAssign package.}
\usage{
AutoAnnotate_CellAssign(SeuratObj, marker_list)
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{marker_list}{a named list, where the names are the cell types and
the entries are marker genes (not necessarily mutually exclusive) for each cell type}
}
\value{
Seurat object. The predicted cell type label is stored in \code{SeuratObj$predicted_label}.
}
\description{
Automatic annotation of Seurat object using CellAssign package.
}
