% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auto_annotation.R
\name{AutoAnnotate_Garnett}
\alias{AutoAnnotate_Garnett}
\title{Automatic annotation of Seurat object using Garnett package.}
\usage{
AutoAnnotate_Garnett(SeuratObj, marker_file_path)
}
\arguments{
\item{SeuratObj}{Seurat object to annotate}

\item{marker_file_path}{see https://cole-trapnell-lab.github.io/garnett/docs/ to construct a marker file}
}
\value{
Seurat object. The predicted cell type label is stored in \code{SeuratObj$predicted_label}.
}
\description{
Automatic annotation of Seurat object using Garnett package.
}
