% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shinyApp.R
\name{load_shinyApp}
\alias{load_shinyApp}
\title{visualize the Seurat object in a built-in shiny app}
\usage{
load_shinyApp(SeuratObj, rasMat, rssMat, tf_target)
}
\arguments{
\item{SeuratObj}{Seurat object}
}
\description{
visualize the Seurat object in a built-in shiny app
}
\examples{
\dontrun{
if(interactive()) {
    load_shinyApp(SeuratObj)
}
}

}
