% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMET_analysis.R
\name{PMEThmp}
\alias{PMEThmp}
\title{Heatmap of PMET result}
\usage{
PMEThmp(PMETresult, topn = NULL)
}
\arguments{
\item{PMETresult}{PMET result}

\item{topn}{show top highly enriched pairs of each cluster}
}
\description{
Heatmap of PMET result
}
