% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGSEA_plant.R
\name{RunGSEA_plant}
\alias{RunGSEA_plant}
\title{Perform Gene Set Enrichment Analysis (GSEA) on plant single-cell data.}
\usage{
RunGSEA_plant(
  SeuratObj,
  by = "GO",
  TERM2GENE = NULL,
  GeneIDtype = "TAIR",
  minpct = 0,
  release_version = "TAIR10"
)
}
\arguments{
\item{SeuratObj}{Seurat object}

\item{by}{GO, KEGG. Will be ignored if parameter "TERM2GENE" is not NULL.}

\item{TERM2GENE}{Customized terms. Annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene. NULL by default. If not NULL, parameter "by" will be ignored.}

\item{GeneIDtype}{Gene ID type in Seurat object. \code{TAIR} for \code{Arabidopsis thaliana}, for example.
This parameter only works when \code{SeuratObj@misc$species == 'Arabidopsis thaliana'}.}

\item{minpct}{Minimum expressed percent in each cluster, used to filter differentially expressed genes to run GSEA. 0 as default.}

\item{release_version}{Release version of genome annotation. 'IRGSP-1.0' or 'v7.0' for Oryza sativa.
'Zm-B73-REFERENCE-NAM-5.0' or 'RefGen_V4' for Zea mays. This parameter will be ignored if \code{SeuratObj@misc$species == 'Arabidopsis thaliana'}.}
}
\value{
Seurat object
}
\description{
Perform Gene Set Enrichment Analysis (GSEA) on Seurat object. GSEA is implemented using \code{clusterProfiler} package.
This function is only for three plant species: \code{Arabidopsis thaliana} or \code{Oryza sativa} or \code{Zea mays}.
}
\examples{
\dontrun{
SeuratObj <- RunGSEA_plant(SeuratObj, by = 'GO')
}

}
