% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRN_visualization.R
\name{toptargets}
\alias{toptargets}
\title{Network diagram showing top targets of each regulon}
\usage{
toptargets(tf_target, topn = 5, regulons = NULL)
}
\arguments{
\item{tf_target}{TF and targets of each regulon. This information comes from post-processing of pyscenic.}

\item{topn}{number of top targets to draw}

\item{regulons}{regulons to show. If NULL, show all regulons.}
}
\description{
Network diagram showing top targets of each regulon according to importance score.
}
